+++ src/components/color-picker/gradient-select/GradientPreview.tsx
